requirement.txt
libusb1
enum34
Pillow
